[elastic-agent-system-metrics] metric/system/cgroup/cgv2: ensure Rdev is correct width
On mips platforms, Stat_t.Rdev is uint32 resulting in a type error. So
explicitly convert to uint64.

+++ metric/system/cgroup/cgv2/io_helper_linux.go
@@ -51,7 +51,7 @@ func fetchDeviceName(major uint64, minor uint64) (bool, string, error) {
 		if !ok {
 			return nil
 		}
-		devID = infoT.Rdev
+		devID = uint64(infoT.Rdev)
 
 		// do some bitmapping to extract the major and minor device values
 		// The odd duplicated logic here is to deal with 32 and 64 bit values.
